+++ Misc/FftInterpolation.cpp
@@ -65,7 +65,7 @@ void FftInterpolation (BaseGrid &G, double *coarse, double *fine, int ratio, boo
   int dimz_c = G.get_PZ0_GRID(1); 
   int dimy_f = G.get_PY0_GRID(ratio); 
   int dimz_f = G.get_PZ0_GRID(ratio); 
-  double scale = 1.0 / (double)(n[0]*n[1]*n[2]);
+  double scale = 1.0 / ((double)n[0]*(double)n[1]*(double)n[2]);
   double dratio = (double)ratio;
   double rootrho;
